created viewModels for each entity and made changed to the firbase to make it work

+++ app/src/main/java/fullstack/application/spherebeat/dal/repository/PlaylistRepository.kt
@@ -1,66 +1,51 @@
-package fullstack.application.spherebeat.repository
+package fullstack.application.spherebeat.dal.repository
 
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import fullstack.application.spherebeat.dal.remote.FirebaseModel
 import fullstack.application.spherebeat.model.Playlist
 import fullstack.application.spherebeat.dal.local.AppLocalDb
 import fullstack.application.spherebeat.dal.local.AppLocalDbRepository
 import java.util.concurrent.Executors
 
-class PlaylistRepository() {
+class PlaylistRepository {
     private var executor = Executors.newSingleThreadExecutor()
     private val firebaseModel: FirebaseModel = FirebaseModel()
     private val localDb: AppLocalDbRepository = AppLocalDb.database
-    private var playlistList: LiveData<List<Playlist>>? = null
-    private val loadingState = MutableLiveData<LoadingState>()
-
-    enum class LoadingState {
-        LOADING,
-        NOT_LOADING
-    }
 
     fun getAllPlaylists(): LiveData<List<Playlist>> {
-        loadingState.value = LoadingState.LOADING
-        if (playlistList == null) {
-            playlistList = localDb.playlistDao().getAllPlaylists()
-            refreshAllPlaylists()
-        }
-        loadingState.value = LoadingState.NOT_LOADING
-        return playlistList!!
+        refreshAllPlaylists()
+        return localDb.playlistDao().getAllPlaylists()
     }
 
     fun getPlaylistById(playlistId: String): LiveData<Playlist> {
-        loadingState.value = LoadingState.LOADING
-        val playlist = localDb.playlistDao().getPlaylistById(playlistId)
-        loadingState.value = LoadingState.NOT_LOADING
-        return playlist
+        refreshAllPlaylists()
+        return localDb.playlistDao().getPlaylistById(playlistId)
     }
 
     fun addPlaylist(playlist: Playlist, callback: (Boolean) -> Unit) {
-        loadingState.value = LoadingState.LOADING
         firebaseModel.addPlaylist(playlist) { success ->
-            refreshAllPlaylists()
-            loadingState.value = LoadingState.NOT_LOADING
+            if (success) {
+                refreshAllPlaylists()
+            }
             callback(success)
         }
     }
 
     fun updatePlaylist(playlist: Playlist, callback: (Boolean) -> Unit) {
-        loadingState.value = LoadingState.LOADING
         firebaseModel.updatePlaylist(playlist) { success ->
-            refreshAllPlaylists()
-            loadingState.value = LoadingState.NOT_LOADING
+            if (success) {
+                refreshAllPlaylists()
+            }
             callback(success)
         }
     }
 
     fun deletePlaylist(playlist: Playlist, callback: (Boolean) -> Unit) {
-        loadingState.value = LoadingState.LOADING
         firebaseModel.deletePlaylist(playlist) { success ->
-            localDb.playlistDao().delete(playlist)
-            refreshAllPlaylists()
-            loadingState.value = LoadingState.NOT_LOADING
+            if (success) {
+                localDb.playlistDao().delete(playlist)
+                refreshAllPlaylists()
+            }
             callback(success)
         }
     }
@@ -71,7 +56,7 @@ class PlaylistRepository() {
             executor.execute {
                 var currentTime = lastUpdated
                 for (playlist in playlists) {
-                    localDb.playlistDao().insertAll(playlist)
+                    localDb.playlistDao().insert(playlist)
                     playlist.lastUpdated?.let {
                         if (currentTime < it) {
                             currentTime = it
